SkypeWeb : MSYS2 build and more flexible installer
The installer is now customizable as to which Pidgin variant it will be targeted
against. This is done with defines passed to makensis. There is now a makefile
for MSYS2 too, which allows for definition of such variant with some variables,
as well as uses git if possible to define the product version.