Don't become connected without valid registratino token
The signed-on event fired before we connected to Skype in cases where
the registration token wasn't valid. This led to clients thinking they
could operated as if the user was logged in (send messages etc.) when
they could not. The status change now lives in do_all_the_things(),
where it only takes place if there is a valid registration.