Fix an issue where a chat was joined but messages couldn't be sent
The message sending code expected a "chatname" attribute to be set on
chats to which message were sent. When the chat-joined event fired, this
attribute was not set. This caused sending messages from a chat-joined
callback to fail.

This commit fixes this issue by attempting to fall back to the name
field on the conversation if the "chatname" attribute is not set.