Keep track of PurpleUtilFetchUrlData and cancel them on logout
Fixes use-after-free bugs when the callback is called after the account
is turned off.

This creates a wrapper function skypeweb_fetch_url_request() which takes
a SkypeWebAccount instead of a PurpleAccount, and stores each return
value in a linked list, sa->url_datas. The callbacks must take care of
removing this entry.

The linked list idea is pretty much copy-pasted from built-in prpls,
main difference is that others don't have a wrapper function.

This diff is rather repetitive, but it's as good as it gets.