skypeweb_got_imagemessage: Pass a copy of the data to the imgstore
purple_imgstore_add_with_id() borrows the reference to the image data,
so passing a copy is needed to avoid double-free when the http request
is cancelled and when the last reference to the imgstore entry is
removed.

In the case of clients like bitlbee which don't keep a reference to the
imgstore entry, this can mean an instant crash.

Fixes #374