Store the last message timestamp of group chats
When re-joining a previously joined chat, skypeweb requests the full message
history from the Skype server. This leads to re-displaying messages that are
already present in the conversation logs stored on the disk, if logging is
enabled.

This commit stores the last message timestamp of all group chats for which a
message has been received. Using the "global" last message timestamp is not
possible, since messages in a given group chat might have appeared between
leaving and re-joining.