Fix "No handler found for resource" on alt login
That error is basically a 404 that you get when you set Accept to json
in the api.skype.com server.

The fix is to grab a new rps URL from skypeweb, it lives in the 'rps'
server (edge.skype.com) and has a slightly different path.

This also removes the 'skyper' md5-based login. It's still referenced in
the javascript of skypeweb, but probably unused because the endpoint is
the same we're hitting here and it's dead.

SOAP login still works with no issues, for both skype usernames and
microsoft accounts (it didn't work with skype usernames when i first
wrote this, but i remember at some point later throwing skype usernames
at the SOAP login and having it work)