Ensure sent message is in XHTML format
It appears that Skype client is very sensitive to type of message that
is being sent. Implicit linkifing through `purple_markup_linkify()` for
some reason adds links with uppercase "<A HREF" which is ignored by
skype client. While Pidgin itself seems to handle it correctly, it is
not the case for all libpurple clients ie BitlBee. Follow other
protocols (ie Jabber) and ensure message is XHTML before sending.