sa->pc safety check in request/timeout callbacks
It mitigates a common cause of segmentation faults due to a freed
SkypeAccount structure which may contain corrupted data already,
still the sa->pc may be NULL as it is freed right before the whole
structure. It does not guarantee the crash prevention though, and does
not solve the root problem, as these callbacks should be removed and
not invoked if an account is disconnected.

The crash backtraces are collected in #650, most of them happened in
`skypeweb_post_or_get`.